requirements.txt
# Trust Regions of Graph Propagation
# Python dependencies with pinned versions for reproducibility
# Tested on Python 3.10

# Core
numpy>=1.21.0,<2.0.0
scipy>=1.7.0,<2.0.0
pandas>=1.3.0,<3.0.0

# Deep Learning
torch>=2.0.0,<3.0.0
torch-geometric>=2.4.0,<3.0.0
# PyG dependencies (install with: pip install torch-scatter torch-sparse -f https://data.pyg.org/whl/torch-2.0.0+cu118.html)

# Visualization
matplotlib>=3.5.0,<4.0.0
seaborn>=0.11.0,<1.0.0

# Statistical Analysis
statsmodels>=0.13.0,<1.0.0
scikit-learn>=1.0.0,<2.0.0

# Data handling
networkx>=2.6.0,<4.0.0
ogb>=1.3.0,<2.0.0  # Open Graph Benchmark

# Utilities
tqdm>=4.62.0,<5.0.0
PyYAML>=6.0,<7.0.0

# Optional: Jupyter support
# jupyter>=1.0.0
# ipykernel>=6.0.0
